# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  description: string;
  content: string;
  date: string;
  readTime: string;
  tags: string[];
  published: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBlogPost {
  title: string;
  slug: string;
  description: string;
  content: string;
  date: string;
  readTime?: string;
  tags?: string[];
  published?: boolean;
}

export interface UpdateBlogPost {
  title?: string;
  slug?: string;
  description?: string;
  content?: string;
  date?: string;
  readTime?: string;
  tags?: string[];
  published?: boolean;
}

export interface DeleteResponse {
  success: boolean;
}

export interface PageItemOverride {
  id: number;
  registryType: string;
  slug: string;
  title?: string | null;
  description?: string | null;
  subtitle?: string | null;
  date?: string | null;
  category?: string | null;
  coverGradient?: string | null;
  coverIcon?: string | null;
  tags?: string[] | null;
  role?: string | null;
  period?: string | null;
  gradient?: string | null;
  icon?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UpdatePageItemOverride {
  title?: string | null;
  description?: string | null;
  subtitle?: string | null;
  date?: string | null;
  category?: string | null;
  coverGradient?: string | null;
  coverIcon?: string | null;
  tags?: string[] | null;
  role?: string | null;
  period?: string | null;
  gradient?: string | null;
  icon?: string | null;
}

export interface ThemeSettings {
  id?: number;
  fontSans?: string | null;
  fontDisplay?: string | null;
  fontMono?: string | null;
  accentColor?: string | null;
  backgroundColor?: string | null;
  textColor?: string | null;
  textSecondaryColor?: string | null;
  borderColor?: string | null;
  createdAt?: string;
  updatedAt?: string;
}

export interface UpdateThemeSettings {
  fontSans?: string | null;
  fontDisplay?: string | null;
  fontMono?: string | null;
  accentColor?: string | null;
  backgroundColor?: string | null;
  textColor?: string | null;
  textSecondaryColor?: string | null;
  borderColor?: string | null;
}

export interface AuthResponse {
  authenticated: boolean;
}

export type ListPageItemOverridesParams = {
  registryType?: string;
};