# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: blog-posts
    description: Blog post operations
  - name: page-item-overrides
    description: Page item override operations
  - name: theme-settings
    description: Theme settings operations
  - name: admin
    description: Admin authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /blog-posts:
    get:
      operationId: listBlogPosts
      tags: [blog-posts]
      summary: List all blog posts
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog-posts]
      summary: Create a blog post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPost"
      responses:
        "201":
          description: Created blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "401":
          description: Unauthorized

  /blog-posts/{slug}:
    get:
      operationId: getBlogPostBySlug
      tags: [blog-posts]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found

  /blog-posts/{id}:
    put:
      operationId: updateBlogPost
      tags: [blog-posts]
      summary: Update a blog post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBlogPost"
      responses:
        "200":
          description: Updated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteBlogPost
      tags: [blog-posts]
      summary: Delete a blog post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /page-item-overrides:
    get:
      operationId: listPageItemOverrides
      tags: [page-item-overrides]
      summary: List page item overrides
      parameters:
        - name: registryType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of overrides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PageItemOverride"

  /page-item-overrides/{registryType}/{slug}:
    put:
      operationId: updatePageItemOverride
      tags: [page-item-overrides]
      summary: Create or update a page item override
      security:
        - bearerAuth: []
      parameters:
        - name: registryType
          in: path
          required: true
          schema:
            type: string
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePageItemOverride"
      responses:
        "200":
          description: Updated override
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageItemOverride"
        "201":
          description: Created override
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageItemOverride"
        "401":
          description: Unauthorized

  /theme-settings:
    get:
      operationId: getThemeSettings
      tags: [theme-settings]
      summary: Get site theme settings
      responses:
        "200":
          description: Theme settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThemeSettings"
    put:
      operationId: updateThemeSettings
      tags: [theme-settings]
      summary: Update site theme settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateThemeSettings"
      responses:
        "200":
          description: Updated theme settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThemeSettings"
        "401":
          description: Unauthorized

  /admin/verify:
    post:
      operationId: verifyAdminToken
      tags: [admin]
      summary: Verify admin token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: string
        date:
          type: string
        readTime:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - slug
        - description
        - content
        - date
        - readTime
        - tags
        - published
        - createdAt
        - updatedAt

    CreateBlogPost:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: string
        date:
          type: string
        readTime:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
      required:
        - title
        - slug
        - description
        - content
        - date

    UpdateBlogPost:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: string
        date:
          type: string
        readTime:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean

    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    PageItemOverride:
      type: object
      properties:
        id:
          type: integer
        registryType:
          type: string
        slug:
          type: string
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        subtitle:
          type: string
          nullable: true
        date:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        coverGradient:
          type: string
          nullable: true
        coverIcon:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        role:
          type: string
          nullable: true
        period:
          type: string
          nullable: true
        gradient:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - registryType
        - slug
        - createdAt
        - updatedAt

    UpdatePageItemOverride:
      type: object
      properties:
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        subtitle:
          type: string
          nullable: true
        date:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        coverGradient:
          type: string
          nullable: true
        coverIcon:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        role:
          type: string
          nullable: true
        period:
          type: string
          nullable: true
        gradient:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true

    ThemeSettings:
      type: object
      properties:
        id:
          type: integer
        fontSans:
          type: string
          nullable: true
        fontDisplay:
          type: string
          nullable: true
        fontMono:
          type: string
          nullable: true
        accentColor:
          type: string
          nullable: true
        backgroundColor:
          type: string
          nullable: true
        textColor:
          type: string
          nullable: true
        textSecondaryColor:
          type: string
          nullable: true
        borderColor:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string

    UpdateThemeSettings:
      type: object
      properties:
        fontSans:
          type: string
          nullable: true
        fontDisplay:
          type: string
          nullable: true
        fontMono:
          type: string
          nullable: true
        accentColor:
          type: string
          nullable: true
        backgroundColor:
          type: string
          nullable: true
        textColor:
          type: string
          nullable: true
        textSecondaryColor:
          type: string
          nullable: true
        borderColor:
          type: string
          nullable: true

    AuthResponse:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated