# lib/api-zod/src/generated/types/themeSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ThemeSettings {
  id?: number;
  fontSans?: string | null;
  fontDisplay?: string | null;
  fontMono?: string | null;
  accentColor?: string | null;
  backgroundColor?: string | null;
  textColor?: string | null;
  textSecondaryColor?: string | null;
  borderColor?: string | null;
  createdAt?: string;
  updatedAt?: string;
}