# lib/api-zod/src/generated/types/updateBlogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateBlogPost {
  title?: string;
  slug?: string;
  description?: string;
  content?: string;
  date?: string;
  readTime?: string;
  tags?: string[];
  published?: boolean;
}