# lib/api-zod/src/generated/types/updatePageItemOverride.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePageItemOverride {
  title?: string | null;
  description?: string | null;
  subtitle?: string | null;
  date?: string | null;
  category?: string | null;
  coverGradient?: string | null;
  coverIcon?: string | null;
  tags?: string[] | null;
  role?: string | null;
  period?: string | null;
  gradient?: string | null;
  icon?: string | null;
}